requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
flasgger==0.9.7.1
Werkzeug==3.0.1
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
requests==2.31.0
Pillow==10.1.0
redis==5.0.1
python-dotenv==1.0.0
